* It's becoming the default in webpack 2.0 (currently in beta)
* It fixes Chrome's source map caching issue (since source maps are evaled as part of app.js)
* It should rebuild faster according to webpack docs: https://webpack.github.io/docs/configuration.html#devtool

Rebuild times after a 1 line change to the same, single file:

|  | Before  | After  |
|---|---|---|
|  1 | 4266ms  | 2967ms  |
|  2 | 2087ms  | 2186ms  |
|  3 | 3164ms  | 1714ms  |
| **Avg** | **3172ms** | **2289ms** |

^There's some pretty big variance here, but ... it did have the lowest ever generation time (1714ms), and the lowest max time (2967ms).